start: Support backing up extracted backup archives for deduplication
One key points of borg is that it can de-duplicate data, however we're
currently just uploading the Hassio backup archives as they are to the
repo making the life of borg hard when it tries to deduplicate them,
because of the presency of compressed archives.

So extract the archives before uploading them. This can be handled
easier in future with recently released import-tar feature.